Moved printing coverage and running tests in to their respective structs
