+++ Parasol/main.swift
@@ -11,24 +11,17 @@ import Foundation
 // TODO: make passing Xcode project as an optional argument
 if let xcodeProject = XcodeProject.findXcodeProjectInCurrentDirectory() {
     if case let .Exists(profdataPath, executablePath) = xcodeProject.coverageDataExists {
-        print(profdataPath)
-        print(executablePath)
-        let task = NSTask()
-        task.launchPath = "/usr/bin/xcrun"
-        task.arguments = ["llvm-cov", "show", "-instr-profile", profdataPath, executablePath]
-        task.launch()
-        task.waitUntilExit()
+        Coverage.showCoverage(profdataPath, executablePath: executablePath)
     } else {
         print("Coverage data has not been generated. Would you like to run tests? (y/n)")
         let input = readLine()
         if input == "y" {
-            let task = NSTask()
-            task.launchPath = "/usr/bin/xcodebuild"
-            task.arguments = ["test", "-project", xcodeProject.name, "-scheme", "Parasol", "-enableCodeCoverage", "YES"]
-            print(task.arguments)
-            task.launch()
-            task.waitUntilExit()
-            print(xcodeProject.coverageDataExists)
+            xcodeProject.runTests()
+            if case let .Exists(profdataPath, executablePath) = xcodeProject.coverageDataExists {
+                Coverage.showCoverage(profdataPath, executablePath: executablePath)
+            } else {
+                print("ERROR: Coverage data was not generated by Xcode.")
+            }
         } else if input == "n" {
             
         }
